+++ Runtime/Services/Factory/ResourcePaths.Platform.cs
@@ -12,6 +12,8 @@ namespace PlatformCore.Services.Factory
 			{
 				public const string ColorStyleLibrary = "UI/ColorStyleLibrary";
 				public const string TextStyleLibrary = "UI/TextStyleLibrary";
+				public const string ElementBackground = "UI/ElementBackground";
+				public const string UICursorView = "UI/UICursorView";
 				public const string UIGlobalNotificationView = "UI/UIGlobalNotificationView";
 				public const string UINotificationsView = "UI/UINotificationsView";
 				public const string UINotificationView = "UI/UINotificationView";